I worked on the two images that you had place holders for: LatrineClosed.png SpeechBubble.png
The door looks decent even with the small resolution but the speech bubble resolution needs to be increased in size to not look pixelated.
I have higher res images of both if you want them.
I was also thinking that the speech bubble could be coming out of the heart-shaped whole of the latrine